move array_namespace type
